Look up accessions and other IDs on Entrez

Passes the provided IDs through Entrez which has the effect of normalising the accepted UIDs, and removing invalid UIDs. For web history lists, this forces results to be freshly downloaded (unlike as_id_list() which can use cached results).

Usage

entrez_validate(id_set, .paginate = 5000L, .path = NULL, .call = current_env())

Arguments

id_set

an id_list object.

.paginate

controls how multiple API requests are used to complete the call. Pagination is performed using the retstart and retmax API parameters. When set to an integer, no more than .pagniate items will be requested per API call. When FALSE or 0, only one API request is sent.

.path

path specification for saving raw responses.

.call

call environment to use in error messages/traces. See rlang::topic-error-call and the call argument of cli::cli_abort(). You only need to specify this in internal helper functions that don't need to be mentioned in error messages.

Value

id_list object

See also